+++ Tests/Resolve-MSBuild.test.ps1
@@ -11,20 +11,15 @@ if (!($ProgramFiles = ${env:ProgramFiles(x86)})) {$ProgramFiles = $env:ProgramFi
 $VS2017 = Test-Path "${env:ProgramFiles(x86)}\Microsoft Visual Studio\2017"
 $VSSetup = Get-Module VSSetup -ListAvailable
 
+function Set-Mock($Name, $Block) {
+	Set-Alias $Name Mock-$Name
+	Set-Content function:\Mock-$Name $Block
+}
+
 function Test-MSBuild([Parameter()]$Path) {
 	if ($Path -notlike '*\MSBuild.exe') {Write-Error "Unexpected path $Path"}
 	if (![System.IO.File]::Exists($Path)) {Write-Error "Missing file $Path"}
 	$Path
-}
-
-$calls = @{}
-function Get-Nothing {
-	$calls.Nothing = 1 + $calls['Nothing']
-}
-
-function Set-Mock($Alias, $Command) {
-	Set-Alias $Alias $Command -Scope 1
-	$calls.Clear()
 }
 
 task test15VSSetup -If $VS2017 {
@@ -35,10 +30,9 @@ task test15VSSetup -If $VS2017 {
 }
 
 task test15Guess -If $VS2017 {
-	Set-Mock Get-MSBuild15VSSetup Get-Nothing
+	. Set-Mock Get-MSBuild15VSSetup {}
 	$r = Resolve-MSBuild 15.0
 	Test-MSBuild $r
-	equals $calls.Nothing 1
 	assert ($r -like '*\15.0\*')
 }
 
@@ -61,10 +55,9 @@ task testAll15 -If $VS2017 {
 }
 
 task testAll14 {
-	Set-Mock Get-MSBuild15 Get-Nothing
+	. Set-Mock Get-MSBuild15 {}
 	$r = Resolve-MSBuild
 	Test-MSBuild $r
-	equals $calls.Nothing 1
 	assert ($r -like '*\14.0\*')
 }
 
@@ -79,10 +72,9 @@ task missingNew {
 }
 
 task missing15 {
-	Set-Mock Get-MSBuild15 Get-Nothing
+	. Set-Mock Get-MSBuild15 {}
 	($r = try {Resolve-MSBuild 15.0} catch {$_})
 	assert (($r | Out-String) -like '*Cannot resolve MSBuild 15.0 : *Resolve-MSBuild.test.ps1:*')
-	equals $calls.Nothing 1
 }
 
 task invalidVersion {
@@ -93,4 +85,126 @@ task invalidVersion {
 task alias-of-Resolve-MSBuild {
 	$r = @(Get-Command Resolve-MSBuild)[0]
 	equals "$($r.CommandType)" Alias
+}
+
+task Get-MSBuild15VSSetup -If $VSSetup {
+    . Set-Mock Select-VSSetupInstance {
+        param($Product)
+        if ($Product -eq '*') {
+            $all
+        }
+        else {
+            foreach ($_ in $all) {
+                if ($_.Product -eq $Product) {
+                    $_
+                }
+            }
+        }
+    }
+
+	$all = @{Product = 'Microsoft.VisualStudio.Product.Enterprise'; InstallationPath = 'Enterprise'}
+	($r = Resolve-MSBuild)
+	equals $r Enterprise\MSBuild\15.0\Bin\MSBuild.exe
+
+	$all = @{Product = 'Microsoft.VisualStudio.Product.Professional'; InstallationPath = 'Professional'}
+	($r = Resolve-MSBuild)
+	equals $r Professional\MSBuild\15.0\Bin\MSBuild.exe
+
+	$all = @{Product = 'Microsoft.VisualStudio.Product.Community'; InstallationPath = 'Community'}
+	($r = Resolve-MSBuild)
+	equals $r Community\MSBuild\15.0\Bin\MSBuild.exe
+
+	$all = @{Product = 'Microsoft.VisualStudio.Product.Something'; InstallationPath = 'Something'}
+	($r = Resolve-MSBuild)
+	equals $r Something\MSBuild\15.0\Bin\MSBuild.exe
+
+	$all = @(
+		@{Product = 'Microsoft.VisualStudio.Product.BuildTools'; InstallationPath = 'BuildTools'}
+		@{Product = 'Microsoft.VisualStudio.Product.Community'; InstallationPath = 'Community'}
+		@{Product = 'Microsoft.VisualStudio.Product.Enterprise'; InstallationPath = 'Enterprise'}
+		@{Product = 'Microsoft.VisualStudio.Product.Professional'; InstallationPath = 'Professional'}
+		@{Product = 'Microsoft.VisualStudio.Product.TeamExplorer'; InstallationPath = 'TeamExplorer'}
+	)
+	($r = Resolve-MSBuild)
+	equals $r Enterprise\MSBuild\15.0\Bin\MSBuild.exe
+
+	$all = @(
+		@{Product = 'Microsoft.VisualStudio.Product.BuildTools'; InstallationPath = 'BuildTools'}
+		@{Product = 'Microsoft.VisualStudio.Product.Community'; InstallationPath = 'Community'}
+		@{Product = 'Microsoft.VisualStudio.Product.Professional'; InstallationPath = 'Professional'}
+		@{Product = 'Microsoft.VisualStudio.Product.TeamExplorer'; InstallationPath = 'TeamExplorer'}
+	)
+	($r = Resolve-MSBuild)
+	equals $r Professional\MSBuild\15.0\Bin\MSBuild.exe
+
+	$all = @(
+		@{Product = 'Microsoft.VisualStudio.Product.BuildTools'; InstallationPath = 'BuildTools'}
+		@{Product = 'Microsoft.VisualStudio.Product.Community'; InstallationPath = 'Community'}
+		@{Product = 'Microsoft.VisualStudio.Product.TeamExplorer'; InstallationPath = 'TeamExplorer'}
+	)
+	($r = Resolve-MSBuild)
+	equals $r Community\MSBuild\15.0\Bin\MSBuild.exe
+
+	$all = @(
+		@{Product = 'Microsoft.VisualStudio.Product.BuildTools'; InstallationPath = 'BuildTools'}
+		@{Product = 'Microsoft.VisualStudio.Product.TeamExplorer'; InstallationPath = 'TeamExplorer'}
+	)
+	($r = Resolve-MSBuild)
+	equals $r BuildTools\MSBuild\15.0\Bin\MSBuild.exe
+}
+
+task Get-MSBuild15Guess {
+	. Set-Mock Get-MSBuild15VSSetup {}
+	. Set-Mock Test-Path {$true}
+	. Set-Mock Resolve-Path {$all}
+
+	$all = @{ProviderPath = '..\Enterprise\..'}
+	($r = Resolve-MSBuild)
+	equals $r '..\Enterprise\..'
+
+	$all = @{ProviderPath = '..\Professional\..'}
+	($r = Resolve-MSBuild)
+	equals $r '..\Professional\..'
+
+	$all = @{ProviderPath = '..\Community\..'}
+	($r = Resolve-MSBuild)
+	equals $r '..\Community\..'
+
+	$all = @{ProviderPath = '..\Something\..'}
+	($r = Resolve-MSBuild)
+	equals $r '..\Something\..'
+
+	$all = @(
+		@{ProviderPath = '..\BuildTools\..'}
+		@{ProviderPath = '..\Community\..'}
+		@{ProviderPath = '..\Enterprise\..'}
+		@{ProviderPath = '..\Professional\..'}
+		@{ProviderPath = '..\TeamExplorer\..'}
+	)
+	($r = Resolve-MSBuild)
+	equals $r '..\Enterprise\..'
+
+	$all = @(
+		@{ProviderPath = '..\BuildTools\..'}
+		@{ProviderPath = '..\Community\..'}
+		@{ProviderPath = '..\Professional\..'}
+		@{ProviderPath = '..\TeamExplorer\..'}
+	)
+	($r = Resolve-MSBuild)
+	equals $r '..\Professional\..'
+
+	$all = @(
+		@{ProviderPath = '..\BuildTools\..'}
+		@{ProviderPath = '..\Community\..'}
+		@{ProviderPath = '..\TeamExplorer\..'}
+	)
+	($r = Resolve-MSBuild)
+	equals $r '..\Community\..'
+
+	$all = @(
+		@{ProviderPath = '..\BuildTools\..'}
+		@{ProviderPath = '..\TeamExplorer\..'}
+	)
+	($r = Resolve-MSBuild)
+	equals $r '..\BuildTools\..'
 }